requirements.txt
torch
git+https://github.com/openai/CLIP.git

faiss-cpu

flask
flask-cors

gradio

fastapi
uvicorn